requirements.txt
openai-whisper
deep-translator
edge-tts
pydub
ffmpeg-python
flask
